# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API: Clients, Products and Orders",
        "description": "API: Client, Product and Orders",
        "contact": {
            "name": "Fernando Marcelino",
            "url": "https://github.com/fernandoomarcelino",
            "email": "ffernandomarcelino@gmail.com"
        },
        "license": {
            "name": "Private",
            "url": "https://github.com/fernandoomarcelino"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "L5 Swagger OpenApi dynamic host server"
        }
    ],
    "paths": {
        "/api/client": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Show all",
                "description": "Show all",
                "operationId": "ClientIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Create",
                "description": "Create",
                "operationId": "ClientStore",
                "requestBody": {
                    "request": "ClientStoreRequest",
                    "description": "ClientStoreRequest",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Missing parameters (validation)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/{id}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Show detail",
                "description": "Show detail",
                "operationId": "ClientShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Client"
                ],
                "summary": "Update",
                "description": "Update",
                "operationId": "ClientUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "ClientUpdateRequest",
                    "description": "ClientUpdateRequest",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Missing parameters (validation)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Remove",
                "description": "Remove",
                "operationId": "ClientDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ClientStoreRequest": {
                "title": "ClientStoreRequest",
                "description": "The request to create Client",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone",
                        "type": "string",
                        "format": "string"
                    },
                    "date_birth": {
                        "title": "date_birth",
                        "description": "date_birth",
                        "type": "string",
                        "format": "string"
                    },
                    "address": {
                        "title": "address",
                        "description": "address",
                        "type": "string",
                        "format": "string"
                    },
                    "complement": {
                        "title": "complement",
                        "description": "complement",
                        "type": "string",
                        "format": "string"
                    },
                    "neighborhood": {
                        "title": "neighborhood",
                        "description": "neighborhood",
                        "type": "string",
                        "format": "string"
                    },
                    "zipcode": {
                        "title": "zipcode",
                        "description": "zipcode",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object"
            },
            "ClientUpdateRequest": {
                "title": "ClientUpdateRequest",
                "description": "The request to update Client"
            },
...
            "ValidationErrorResponse": {
                "title": "Validation Error response",
                "description": "The validation error response",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "error message",
                        "format": "string"
                    },
                    "errors": {
                        "title": "errors",
                        "description": "validation errors in the format: Property name: validation error",
                        "format": "array",
                        "items": {
                            "properties": {
                                "field": {
                                    "title": "field",
                                    "format": "string"
                                },
                                "value": {
                                    "title": "value",
                                    "format": "string"
                                }
                            },
                            "type": "object",
                            "format": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
    "security": [
        []
    ]
}